Getting close by separating UI update from network logic

+++ InetSpeedTesting/Datum.cpp
@@ -1,3 +1,3 @@
 #include "pch.h"
 #include "Datum.h"
-InetSpeedTesting::Datum::Datum(uint16 num, float32 rtt, float32 download) :TestNum(num), RTT(rtt), DownloadBW(download){}
+InetSpeedTesting::Datum::Datum(uint32 num, float32 rtt, float32 download) :TestNum(num), RTT(rtt), DownloadBW(download){}
